add: state.rs
Separated StatefulDir from main.rs and made it state.rs.
